requirements.txt
cachetools==4.2.2
celery==5.1.2  #
django-authtools==1.7.0
django-cache-memoize==0.1.10
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-db-geventpool==4.0.0
django-debug-toolbar
django-debug-toolbar-force
django-environ==0.4.5
django-filter==2.4.0
django-model-utils==4.1.1
django-redis==5.0.0
django-s3-storage==0.13.4
django-timezone-field==4.2.1
django==3.2.5
djangorestframework-camel-case==1.2.0
djangorestframework==3.12.4
docutils==0.17.1
drf-yasg[validation]==1.20.0
ethereum==2.3.2
firebase-admin==5.0.1
gnosis-yodaplus-py[django]==3.1.16
gunicorn[gevent]==20.1.0
hexbytes==0.2.1
packaging>=21.0
pillow==8.3.1
psycogreen==1.0.2
psycopg2-binary==2.9.1
redis==3.5.3
requests==2.26.0
web3==5.19.0
